s3/client: add upload_ranged_sink
this sink prepares for the backup feature, where the object to be
stored in the object storage is known beforehand. with the predicatable
total size and write pattern of the sink, we don't need to accumulate
the payloads in memory and send them in batch, as we do in upload_sink
and in upload_jumbo_sink. this puts less pressure on the memory
subsystem.

so, a new upload_sink is implemented. it is supposed to used when
we upload an sstable from local filesystem to the object storage.

at first glance at the `test_client_multipart_upload_ranged` test,
upload_ranged_sink comes with harsh constraints. but please note,  in
Seastar, a `data_sink` is always used along with an `output_stream`,
which has its own zero-copy buffer. we don't have to keep a buffer in
sink, because the `output_stream` on top of it has its own buffer. and
we are able to control its behavior with proper settings when
contructing an `output_stream`.

but to accomodate the needs of `output_stream` which does not
call `put()` at with the expected packet / buffer. `upload_ranged_sink`
also comes with its own buffer. but with the optimal settings, its
buffer is only used for converting `net::packet` to `memory_data_sink_buffers`.

Fixes #16287
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>