tests: tablets: Do read barrier in get_tablet_replicas()
In order for the call to see all prior changes to group0. Also, we
should query on the host on which we executed the barrier.

I hope this will reduce flakiness observed in CI runs on
https://github.com/scylladb/scylladb/pull/16341 where the expected
tablet replica didn't match the one returned by get_tablet_replica()
after tablet movement, possibly because the node is still behind
group0 changes.